Refactor app menu to be less of a monolith
