+++ qubes_menu/utils.py
@@ -1,0 +1,51 @@
+# -*- encoding: utf8 -*-
+#
+# The Qubes OS Project, http://www.qubes-os.org
+#
+# Copyright (C) 2021 Marta Marczykowska-Górecka
+#                               <marmarta@invisiblethingslab.com>
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU Lesser General Public License as published by
+# the Free Software Foundation; either version 2.1 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License along
+# with this program; if not, see <http://www.gnu.org/licenses/>.
+
+# pylint: disable=wrong-import-position
+import gi
+gi.require_version('Gtk', '3.0')
+from gi.repository import Gtk, GdkPixbuf, GLib
+
+
+def load_icon(icon_name, size: Gtk.IconSize = Gtk.IconSize.LARGE_TOOLBAR):
+    _, width, height = Gtk.icon_size_lookup(size)
+    try:
+        return GdkPixbuf.Pixbuf.new_from_file_at_size(icon_name, width, height)
+    except (GLib.Error, TypeError):
+        try:
+            # icon name is a path
+            image: GdkPixbuf.Pixbuf = Gtk.IconTheme.get_default().load_icon(
+                icon_name, width, 0)
+            return image
+        except TypeError:
+            # icon not found in any way
+            pixbuf: GdkPixbuf.Pixbuf = GdkPixbuf.Pixbuf.new(
+                GdkPixbuf.Colorspace.RGB, True, 8, width, height)
+            pixbuf.fill(0x000)
+            return pixbuf
+
+
+def show_error(title, text):
+    dialog = Gtk.MessageDialog(
+        None, 0, Gtk.MessageType.ERROR, Gtk.ButtonsType.OK)
+    dialog.set_title(title)
+    dialog.set_markup(text)
+    dialog.connect("response", lambda *x: dialog.destroy())
+    dialog.show()
