+++ rpm_spec/qubes-desktop-linux-menu.spec.in
@@ -97,6 +97,15 @@ gtk-update-icon-cache %{_datadir}/icons/Adwaita &>/dev/null || :
 %{python3_sitelib}/qubes_menu/__pycache__/*
 %{python3_sitelib}/qubes_menu/__init__.py
 %{python3_sitelib}/qubes_menu/appmenu.py
+%{python3_sitelib}/qubes_menu/app_widgets.py
+%{python3_sitelib}/qubes_menu/application_page.py
+%{python3_sitelib}/qubes_menu/constants.py
+%{python3_sitelib}/qubes_menu/custom_widgets.py
+%{python3_sitelib}/qubes_menu/desktop_file_manager.py
+%{python3_sitelib}/qubes_menu/favorites_page.py
+%{python3_sitelib}/qubes_menu/settings_page.py
+%{python3_sitelib}/qubes_menu/utils.py
+%{python3_sitelib}/qubes_menu/vm_manager.py
 %{python3_sitelib}/qubes_menu/qubes-menu.glade
 %{python3_sitelib}/qubes_menu/qubes-menu-dark.css
